fix: Round H — unchecked indexing, lock poison, expect elimination, overflow guards
MEDIUM:
- phi_engine: n_nodes underflow in ExhaustivePartition time estimate →
  saturating_sub + saturating_pow
- hierarchical_ltc: div-by-zero on empty circuits → early return guard
- STT batch.rs: 19× .lock().unwrap() → poison-recovery pattern
  (unwrap_or_else(|e| e.into_inner()))
- sleep_sentinel: 3× .expect() on fallible LTC ops → graceful
  zero-output degradation

LOW:
- adaptive_reasoning: empty-slice panic → fallback guard
- qwen3: unreachable!() → descriptive message
- phi_engine: 3× unreachable!() → descriptive messages
- video/source.rs: u32 width*height*3 overflow → usize arithmetic
- fractal.rs: branching.pow() overflow → checked_pow
- dual_layer MockTransport: 7× lock poison → recovery pattern

9 files, 49 insertions, 37 deletions. 0 regressions (3954 pass,
3 pre-existing CfC tau failures).

+++ src/perception/video/source.rs
@@ -162,7 +162,7 @@ impl VideoSource for WebcamSource {
         // Generate synthetic frame data with some temporal variation
         let width = self.config.width;
         let height = self.config.height;
-        let size = (width * height * 3) as usize;
+        let size = (width as usize) * (height as usize) * 3;
 
         // Create a pattern that changes over time (for testing temporal detection)
         let phase = (self.sequence as f32 * 0.1).sin();
@@ -299,7 +299,7 @@ impl MockVideoSource {
     fn generate_frame(&mut self) -> Frame {
         let width = self.config.width;
         let height = self.config.height;
-        let size = (width * height * 3) as usize;
+        let size = (width as usize) * (height as usize) * 3;
         let seq = self.sequence;
 
         let fps = self.config.fps as f32;
